+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KQYP08CGYGJ172VTW92M6BBD"
+        "01KR0ABMN14XAJM054KT55H5BX"
     )
     r"""The unique identifier of the data source"""