+++ packages/orq-rc/src/orq_ai_sdk/models/updateidentityop.py
@@ -194,7 +194,7 @@ class UpdateIdentityResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-05-06T13:00:13.358Z")
+    updated: Optional[datetime] = parse_datetime("2026-05-07T04:15:13.475Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")