+++ packages/orq-rc/RELEASES.md
@@ -7609,3 +7609,13 @@ Based on:
 - [python v4.9.0-rc.28] packages/orq-rc
 ### Releases
 - [PyPI v4.9.0-rc.28] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.28 - packages/orq-rc
+
+## 2026-05-07 04:18:15
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.762.0 (2.882.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.9.0-rc.29] packages/orq-rc
+### Releases
+- [PyPI v4.9.0-rc.29] https://pypi.org/project/orq-ai-sdk/4.9.0-rc.29 - packages/orq-rc